+++ apps/client/src/app/pages/portfolio/activities/create-or-update-activity-dialog/interfaces/interfaces.ts
@@ -2,7 +2,7 @@ import { Activity } from '@ghostfolio/api/app/order/interfaces/activities.interf
 import { User } from '@ghostfolio/common/interfaces';
 import { Account } from '@prisma/client';
 
-export interface CreateOrUpdateTransactionDialogParams {
+export interface CreateOrUpdateActivityDialogParams {
   accountId: string;
   accounts: Account[];
   activity: Activity;